Updated the Design
Now the Main-Programm is a loop of two files.

checkSettings reads the Informations out of the config.
It also creates a new config.txt if it isnt found.
When all settings are extracted, it passes them to the main function.

The Main function does all the work on checking and moving the files. After the Main is done, it calls the checkSettings function again to update, if the settings have been changed.