Changed access testing
The old Version did not work.
I now tried a Version where I rename the File to the same Name. If the File is opened, I theoreticly shouldnt be able to rename it. At least with a .txt and a .pdf it worked.
If the File cant be renamed, continue and try again later.

The Mail Stuff isnt needed anymore.